Add files via upload
The team are building a model that requires GDP data at a monthly granularity. The Office
of National Statistics publish Quarterly GDP data online in a CSV format to the following
schedule:

. December GDP is published in February.
. March GDP is published in May.
. June GDP is published in August.
. September GDP is published in November.

Please see attached the August 2024 extract from the ONS. It is assumed that the GDP will
have increased by 4.5% from December 2023 to December 2024. Further to this, annual
percentage change assumptions for the following year will be made towards the end of each
year - the 2025 assumption is a 4.7% increase from December 2024 to December 2025. It is
also assumed that GDP changes linearly between quarters.